requirements.txt
fastapi[all]
fastapi-users[sqlalchemy]
sqlalchemy
asyncpg
alembic
python-dotenv
SQLAlchemy-ImageAttach
SQLAlchemy-Utils
py-fastapi-logging